requirements.txt
openai-whisper>=20231117
numpy>=1.23
scikit-learn>=1.2
tqdm>=4.66
pytest>=8.0
